Update libwav.h
Use my new code style.

+++ libwav.h
@@ -54,29 +54,29 @@ struct _wav_file
 {
 	wav_header header;
 	wav_format format;
-	unsigned int datablocks;
+	size_t datablocks;
 	int *data;
 };
 
 typedef struct _wav_file wav_file;
 
-/* WAV_FORMAT */
+// WAV_FORMAT
 enum wav_error wav_format_write (const wav_format*, FILE*);
 enum wav_error wav_format_read (wav_format*, FILE*);
 void wav_format_print (const wav_format*);
 
-/* WAV_HEADER */
+// WAV_HEADER
 enum wav_error wav_header_write (const wav_header*, FILE*);
 enum wav_error wav_header_read (wav_header*, FILE*);
 void wav_header_print (const wav_header*);
 
-/* WAV_CHUNK */
+// WAV_CHUNK
 void wav_chunk_init (wav_chunk*, const char*, const unsigned int, const void*);
 enum wav_error wav_chunk_write (const wav_chunk*, FILE*);
 enum wav_error wav_chunk_read (wav_chunk*, FILE*);
 void wav_chunk_print (const wav_chunk*);
 
-/* WAV_FILE */
+// WAV_FILE
 void wav_free (wav_file*);
 enum wav_error wav_write (const wav_file*, const char*);
 enum wav_error wav_read (wav_file*, const char*);
